+++ lib/ui/widgets/draggable_scrollbar.dart
@@ -64,8 +64,7 @@ class DraggableScrollbarState extends State<DraggableScrollbar> {
 
   // if list takes 300.0 pixels of height on screen and scrollthumb height is 40.0
   // then max bar offset is 260.0
-  double get barMaxScrollExtent =>
-      (context.size!.height - (widget.scrollbarHeight + widget.scrollbarBottomMargin)).abs();
+  double get barMaxScrollExtent => (context.size!.height - (widget.scrollbarHeight + widget.scrollbarBottomMargin)).abs();
   double get barMinScrollExtent => widget.scrollbarTopMargin;
 
   // this is usually length (in pixels) of list
@@ -291,7 +290,7 @@ class DraggableScrollbarState extends State<DraggableScrollbar> {
           _barOffsetBottom = 0;
         }
 
-      // glue scroll bar to the top on overscroll (android only):
+        // glue scroll bar to the top on overscroll (android only):
       } else if (notification is OverscrollNotification) {
         if (_barOffsetTop > barMinScrollExtent && notification.overscroll < 0) {
           setState(() {
@@ -299,7 +298,7 @@ class DraggableScrollbarState extends State<DraggableScrollbar> {
           });
         }
       }
-      
+
       /*else if (notification is OverscrollNotification) {
         var diff = viewMaxScrollExtent - _viewOffset;
         if (diff != 0 && diff < 300) {
@@ -349,8 +348,7 @@ class DraggableScrollbarState extends State<DraggableScrollbar> {
                 width: widget.scrollbarInvisibleWidth,
                 height: widget.scrollbarHeight,
                 // padding: EdgeInsets.only(left: 20.0, top: 10.0, bottom: 10.0),
-                margin: EdgeInsets.only(
-                    top: _barOffsetTop >= 0 ? _barOffsetTop : 0, bottom: _barOffsetBottom >= 0 ? _barOffsetBottom : 0),
+                margin: EdgeInsets.only(top: _barOffsetTop >= 0 ? _barOffsetTop : 0, bottom: _barOffsetBottom >= 0 ? _barOffsetBottom : 0),
                 child: _buildScrollThumb(),
               ),
             ),
@@ -369,9 +367,7 @@ class DraggableScrollbarState extends State<DraggableScrollbar> {
       decoration: BoxDecoration(
         // borderRadius: BorderRadius.only(topLeft: Radius.circular(10.0), bottomLeft: Radius.circular(10.0)),
         // borderRadius: BorderRadius.all(Radius.circular(10)),
-        borderRadius: _isDragInProcess
-            ? const BorderRadius.all(Radius.circular(10))
-            : const BorderRadius.all(Radius.circular(1.5)),
+        borderRadius: _isDragInProcess ? const BorderRadius.all(Radius.circular(10)) : const BorderRadius.all(Radius.circular(1.5)),
         color: widget.scrollbarColor,
       ),
     );